+++ examples/examples_test.go
@@ -15,10 +15,13 @@
 package examples
 
 import (
+	"io/ioutil"
+	"net/http"
 	"os"
 	"path"
 	"path/filepath"
 	"testing"
+	"time"
 
 	"github.com/stretchr/testify/assert"
 
@@ -75,6 +78,21 @@ func TestExamples(t *testing.T) {
 			// One change is known to occur during refresh of the resources in this example:
 			// * `~  gcp:storage:Bucket f-bucket updated changes: + websites`
 			ExpectRefreshChanges: true,
+			ExtraRuntimeValidation: validateAPITest(func(body string) {
+				assert.Equal(t, "Hello World!", body)
+			}),
+		}),
+		jsBase.With(integration.ProgramTestOptions{
+			Dir: path.Join(cwd, "bucket"),
+			// One change is known to occur during refresh of the resources in this example:
+			// * `~  gcp:storage:Bucket f-bucket updated changes: + websites`
+			ExpectRefreshChanges: true,
+		}),
+		jsBase.With(integration.ProgramTestOptions{
+			Dir: path.Join(cwd, "topic"),
+			// One change is known to occur during refresh of the resources in this example:
+			// * `~  gcp:storage:Bucket f-bucket updated changes: + websites`
+			ExpectRefreshChanges: true,
 		}),
 		pythonBase.With(integration.ProgramTestOptions{
 			Dir: path.Join(cwd, "minimal-py"),
@@ -111,4 +129,27 @@ func TestExamples(t *testing.T) {
 
 func createEditDir(dir string) integration.EditDir {
 	return integration.EditDir{Dir: dir, ExtraRuntimeValidation: nil}
+}
+
+func validateAPITest(isValid func(body string)) func(t *testing.T, stack integration.RuntimeValidationStackInfo) {
+	return func(t *testing.T, stack integration.RuntimeValidationStackInfo) {
+		var resp *http.Response
+		var err error
+		url := stack.Outputs["url"].(string)
+		// Retry a couple times on 5xx
+		for i := 0; i <= 5; i++ {
+			resp, err = http.Get(url)
+			if !assert.NoError(t, err) {
+				return
+			}
+			if resp.StatusCode < 500 {
+				break
+			}
+			time.Sleep(1 * time.Minute)
+		}
+		defer resp.Body.Close()
+		body, err := ioutil.ReadAll(resp.Body)
+		assert.NoError(t, err)
+		isValid(string(body))
+	}
 }