+++ sdk/nodejs/storage/index.ts
@@ -18,3 +18,4 @@ export * from "./notification";
 export * from "./objectACL";
 export * from "./objectAccessControl";
 export * from "./transferJob";
+export * from "./zMixins";