+++ CHANGELOG.md
@@ -4,6 +4,16 @@
 
 - Update to v2.3.0 of the Google Beta Terraform Provider.
 
+- Adds a new gcp.cloudfunctions.CallbackFunction component.  This component makes it simple to
+  create Cloud Functions from an existing JavaScript function.  See
+  https://github.com/pulumi/docs/blob/master/reference/serializing-functions.md for more details.
+
+- Using the above, serverless event handlers have been added that enable creating Cloud Functions
+  from a JavaScript function and then invoking that function in response to:
+    1. http triggers.  i.e. when accessing some url.
+    2. storage triggers.  For example, when an object is created in a bucket.
+    3. pubsub triggers.  For example, when a message is published to a topic.
+
 ## 0.18.0 (Released March 21, 2019)
 
 ### Important